+++ libs/core/dotnet/application/Commands/CommandDefinition.cs
@@ -4,7 +4,7 @@ namespace OpenSystem.Core.Application.Commands
 {
     public class CommandDefinition : VersionDefinition
     {
-        public CommandDefinition(uint version, Type type, string name)
+        public CommandDefinition(ulong version, Type type, string name)
             : base(version, type, name) { }
     }
 }
