+++ libs/core/dotnet/domain/Attributes/EventVersionAttribute.cs
@@ -5,7 +5,7 @@ namespace OpenSystem.Core.Domain.Attributes
     [AttributeUsage(AttributeTargets.Class, AllowMultiple = true)]
     public class EventVersionAttribute : VersionAttribute
     {
-        public EventVersionAttribute(string name, uint version)
+        public EventVersionAttribute(string name, ulong version)
             : base(name, version) { }
     }
 }
