+++ libs/core/dotnet/domain/Attributes/JobVersionAttribute.cs
@@ -5,7 +5,7 @@ namespace OpenSystem.Core.Domain.Attributes
     [AttributeUsage(AttributeTargets.Class)]
     public class JobVersionAttribute : VersionAttribute
     {
-        public JobVersionAttribute(string name, uint version)
+        public JobVersionAttribute(string name, ulong version)
             : base(name, version) { }
     }
 }
