+++ libs/core/dotnet/domain/Events/EventDefinition.cs
@@ -4,7 +4,7 @@ namespace OpenSystem.Core.Domain.Events
 {
     public class EventDefinition : VersionDefinition
     {
-        public EventDefinition(uint version, Type type, string name)
+        public EventDefinition(ulong version, Type type, string name)
             : base(version, type, name) { }
     }
 }
